requirements.txt
uvicorn
requests
fastapi
python-dotenv
aiofiles
jinja2
beautifulsoup4
deta